# add support for property descriptors