--- a/lib/assign-properties.js
+++ b/lib/assign-properties.js
@@ -1,4 +1,6 @@
+require('./get-own-property-descriptors');
 var deprecation = require('./deprecation');
+
 
 function setPrototypeOf(obj, proto) {
   Object.setPrototypeOf ? Object.setPrototypeOf(obj, proto) : obj.__proto__ = proto;
@@ -6,6 +8,14 @@ function setPrototypeOf(obj, proto) {
 
 function getPrototypeOf(obj, proto) {
   return Object.getPrototypeOf ? Object.getPrototypeOf(obj) : obj.__proto__;
+}
+
+function hasGetter(descriptor) {
+  return descriptor.hasOwnProperty('get') && descriptor.get !== undefined;
+}
+
+function hasSetter(descriptor) {
+  return descriptor.hasOwnProperty('set') && descriptor.set !== undefined;
 }
 
 function giveMethodSuper(superclass, name, fn) {
@@ -54,17 +64,42 @@ if (sourceAvailable) {
   }
 }
 
+function mergeDescriptors(descriptor, targetDescriptor) {
+  if (descriptor.hasOwnProperty('value') || !targetDescriptor) {
+    return descriptor;
+  }
+
+  if ( !hasGetter(descriptor) && hasGetter(targetDescriptor) ) {
+    descriptor.get = targetDescriptor.get;
+  }
+
+  if ( !hasSetter(descriptor) && hasSetter(targetDescriptor) ) {
+    descriptor.set = targetDescriptor.set;
+  }
+
+  return descriptor;
+}
+
 function assignProperties(target, options) {
-  var value;
+  var optionDescriptors  = Object.getOwnPropertyDescriptors(options),
+      targetDescriptors  = Object.getOwnPropertyDescriptors(getPrototypeOf(target)),
+      propertiesToDefine = {},
+      value, descriptor, targetDescriptor;
 
   for (var key in options) {
-    value = options[key];
+    descriptor = optionDescriptors[key];
+    value = descriptor.value;
 
     if (typeof value === 'function' && hasSuper(value)) {
       target[key] = giveMethodSuper(getPrototypeOf(target), key, value);
     } else {
-      target[key] = value;
+      targetDescriptor = targetDescriptors[key];
+      propertiesToDefine[key] = mergeDescriptors(descriptor, targetDescriptor);
     }
+  }
+
+  if (Object.keys(propertiesToDefine).length) {
+    Object.defineProperties(target, propertiesToDefine);
   }
 }