@bryanaka 2016-06-25
Ah, actually this didn't resolve the case of passing in getters and setters at instantiation time.

Will add a test w/ code when I get some downtime.