@stefanpenner 2016-06-27
> Ah, actually this didn't resolve the case of passing in getters and setters at instantiation time.

I'm unsure if we actually want to do this. Or can you provide a detailed example. I can totally see us wanting to support accessors at extend time though.