@bryanaka 2016-07-02
I am in favor of not supporting getters and setters at instantiation time. I think in any OO language, this would either be either discouraged or not/poorly supported.

I figure the same argument for this was probably made in some form for not allowing the definition of computed properties on `Ember.Object.create()`