@bryanaka 2016-07-02
@stefanpenner Added `Object.getOwnPropertyDescriptors` with [TC-39's proposal](https://github.com/tc39/proposal-object-getownpropertydescriptors) polyfill with one change (`Reflect.ownKeys => Object.keys`).

Also coalesced the property definition to one call to `Object.defineProperties` which should help a little too.