@stefanpenner 2016-07-11
> I am in favor of not supporting getters and setters at instantiation time. I think in any OO language, this would either be either discouraged or not/poorly supported.

sounds good.

> I figure a similar argument could be made against that behavior that was probably made in some form for not allowing the definition of computed properties on Ember.Object.create()

yup (it as perf related)



> @stefanpenner Added Object.getOwnPropertyDescriptors with TC-39's proposal polyfill with one change (Reflect.ownKeys => Object.keys).

looks good.