@stefanpenner 2016-07-11
@bryanaka mind running the benchmarks again with the recent changes?

I'm off vacation so my responses should be more timely now