@bryanaka 2016-08-04
@stefanpenner sorry got this just as I was going on a vacation - now I am back :)

Be aware that the numbers are going to look quite different since I am on a different machine than the previous benchmarks I ran.

Before:
```
core-object/create (0)  x 2,194,521 ops/sec ±2.26% (75 runs sampled)
core-object/create (1)  x 1,979,167 ops/sec ±1.27% (80 runs sampled)
core-object/create (5)  x 1,789,926 ops/sec ±4.53% (76 runs sampled)
core-object/create (default init) (0)  x 40,111,608 ops/sec ±4.46% (73 runs sampled)
core-object/create (default init) (1)  x 16,172,184 ops/sec ±5.07% (69 runs sampled)
core-object/create (default init) (5)  x 4,940,407 ops/sec ±15.62% (61 runs sampled)
uberproto/create (0)  x 2,762,460 ops/sec ±6.32% (76 runs sampled)
uberproto/create (1)  x 2,966,778 ops/sec ±1.55% (80 runs sampled)
uberproto/create (5)  x 2,988,660 ops/sec ±1.49% (72 runs sampled)
raw/create (0)  x 28,326,307 ops/sec ±1.42% (76 runs sampled)
raw/create (1)  x 32,401,993 ops/sec ±1.56% (79 runs sampled)
raw/create (5)  x 26,712,653 ops/sec ±1.53% (78 runs sampled)
esnext/create (0)  x 12,632,076 ops/sec ±75.00% (45 runs sampled)
esnext/create (1)  x 33,228,345 ops/sec ±1.68% (79 runs sampled)
esnext/create (5)  x 30,233,392 ops/sec ±1.79% (75 runs sampled)
klass/create (0)  x 20,560,784 ops/sec ±1.64% (77 runs sampled)
klass/create (1)  x 19,840,010 ops/sec ±1.66% (82 runs sampled)
klass/create (5)  x 18,764,717 ops/sec ±1.91% (77 runs sampled)
backbone-metal/create (0)  x 3,957,427 ops/sec ±30.41% (50 runs sampled)
backbone-metal/create (1)  x 4,404,259 ops/sec ±5.18% (75 runs sampled)
backbone-metal/create (5)  x 4,090,520 ops/sec ±5.78% (72 runs sampled)
core-object/extend (1)  x 392,657 ops/sec ±5.90% (64 runs sampled)
core-object/extend (5)  x 140,637 ops/sec ±3.83% (67 runs sampled)
uberproto/extend (1)  x 104,112 ops/sec ±4.81% (78 runs sampled)
uberproto/extend (5)  x 37,244 ops/sec ±1.48% (83 runs sampled)
klass/extend (1)  x 79,223 ops/sec ±3.69% (73 runs sampled)
klass/extend (5)  x 27,151 ops/sec ±4.25% (71 runs sampled)
backbone-metal/extend (1)  x 59,234 ops/sec ±3.63% (69 runs sampled)
backbone-metal/extend (5)  x 33,804 ops/sec ±3.08% (75 runs sampled)
```

After:
```
core-object/create (0)  x 2,085,642 ops/sec ±2.36% (76 runs sampled)
core-object/create (1)  x 2,002,813 ops/sec ±1.56% (80 runs sampled)
core-object/create (5)  x 1,937,289 ops/sec ±1.91% (73 runs sampled)
core-object/create (default init) (0)  x 44,447,911 ops/sec ±1.31% (77 runs sampled)
core-object/create (default init) (1)  x 18,901,320 ops/sec ±1.08% (84 runs sampled)
core-object/create (default init) (5)  x 5,929,880 ops/sec ±1.30% (79 runs sampled)
uberproto/create (0)  x 3,122,887 ops/sec ±1.28% (82 runs sampled)
uberproto/create (1)  x 3,003,756 ops/sec ±1.45% (82 runs sampled)
uberproto/create (5)  x 2,942,867 ops/sec ±1.48% (76 runs sampled)
raw/create (0)  x 27,412,096 ops/sec ±1.67% (83 runs sampled)
raw/create (1)  x 32,391,986 ops/sec ±1.66% (74 runs sampled)
raw/create (5)  x 23,358,272 ops/sec ±5.24% (74 runs sampled)
esnext/create (0)  x 25,771,558 ops/sec ±1.82% (77 runs sampled)
esnext/create (1)  x 32,444,258 ops/sec ±2.07% (78 runs sampled)
esnext/create (5)  x 30,056,537 ops/sec ±1.66% (75 runs sampled)
klass/create (0)  x 19,910,491 ops/sec ±1.48% (79 runs sampled)
klass/create (1)  x 20,207,103 ops/sec ±1.55% (74 runs sampled)
klass/create (5)  x 20,060,888 ops/sec ±1.44% (77 runs sampled)
backbone-metal/create (0)  x 5,704,667 ops/sec ±5.61% (70 runs sampled)
backbone-metal/create (1)  x 4,692,624 ops/sec ±5.04% (75 runs sampled)
backbone-metal/create (5)  x 4,415,965 ops/sec ±5.30% (72 runs sampled)
core-object/extend (1)  x 45,392 ops/sec ±2.23% (79 runs sampled)
core-object/extend (5)  x 25,195 ops/sec ±3.70% (65 runs sampled)
uberproto/extend (1)  x 113,612 ops/sec ±1.61% (74 runs sampled)
uberproto/extend (5)  x 36,785 ops/sec ±1.86% (79 runs sampled)
klass/extend (1)  x 83,610 ops/sec ±3.78% (71 runs sampled)
klass/extend (5)  x 29,501 ops/sec ±3.73% (76 runs sampled)
backbone-metal/extend (1)  x 66,079 ops/sec ±3.42% (70 runs sampled)
backbone-metal/extend (5)  x 36,181 ops/sec ±2.72% (73 runs sampled)
```

Unfortunately `CoreObject#extend` still drops pretty significantly.