@bryanaka 2016-08-26
ping @stefanpenner 

I know you are busy, just want to make sure if you'd like any further work done on this PR, I can start on it.