@stefanpenner 2017-01-05
this seems like a vary reasonable addition! Will investigate (and try not to forget)