@stefanpenner 2017-01-05
This does not appear to support `super` in descriptors.

Now that node 0.12 is dropped. One can actually use ES6 classes and CoreObject. I believe that __may__ help?

the following i believe works:

```js
class Foo extends CoreObject {
  get name() { return 'hi'; }
}