@stefanpenner 2017-01-13
As we now have legit class syntax + super (node 0.12 is no longer supported), and that syntax supports accessors by default. I suspect we can close this.

Will gladly reopen if I am mistaken.