@bryanaka 2017-01-13
@stefanpenner I'm all for that. Less code == win in my book almost always

One thing though is that if we want to recommend that usage when needing descriptors, the `super` call syntax changes depending on if you use `CoreObject.extend` or `class Foo extends CoreObject` which is *a little* weird.

Would love to unite that concept but would be a huge breaking change just to get `this._super` to act like `super` and don't think descriptors are mainstream enough to justify that kind of breaking change ATM. Just food for thought to include in #29